mindnotesBackend/requirements.txt
asgiref==3.7.2
Django==3.2.25
djangorestframework==3.14.0
dotenv==0.9.9
pillow==10.0.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
sqlparse==0.2.4
typing_extensions==4.7.1
pymongo==4.6.0
mongoengine==0.27.0
djongo==1.3.6